Add public models for Virksomhetsregisteret
These models is private in a seperate project owned by a separate teams.
Ideally, these models should be as a separate public package, e.g.
Finanstilsynet.Virksomhetsregisteret.Models, available on nuget. We
copy/paste them into this project temporarily. When the models are made
available as a separate package, these should be deleted and code should
be migrated to use those models.